[REFACTOR] separate get methods tests for invoice controller
